requirements.txt
flask
requests
PyJWT
python-dotenv
flasgger
pytest
pytest-cov
pylint
black